rename package to noindex
