+++ noindex/parallel/slice.go
