+++ noindex/parallel/slice_test.go
