+++ noindex/slice_test.go
@@ -1,4 +1,4 @@
-package simple
+package noindex
 
 import (
 	"strconv"
